+++ dev-lang/python/python-3.1.1-r1.ebuild
@@ -1,6 +1,6 @@
 # Copyright 1999-2009 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/dev-lang/python/python-3.1.1-r1.ebuild,v 1.7 2009/09/29 20:02:56 arfrever Exp $
+# $Header: /var/cvsroot/gentoo-x86/dev-lang/python/python-3.1.1-r1.ebuild,v 1.8 2009/10/01 20:03:52 arfrever Exp $
 
 EAPI="2"
 
@@ -40,9 +40,9 @@ RDEPEND=">=app-admin/eselect-python-20090606
 			tk? ( >=dev-lang/tk-8.0 )
 			xml? ( >=dev-libs/expat-2 )
 		)
-		!m68k? ( !mips? ( !sparc-fbsd? ( virtual/libffi ) ) )"
+		!m68k? ( !sparc-fbsd? ( virtual/libffi ) )"
 DEPEND="${RDEPEND}
-		!m68k? ( !mips? ( !sparc-fbsd? ( dev-util/pkgconfig ) ) )"
+		!m68k? ( !sparc-fbsd? ( dev-util/pkgconfig ) )"
 RDEPEND+=" !build? ( app-misc/mime-types )"
 PDEPEND="app-admin/python-updater
 		=dev-lang/python-2*"
@@ -51,7 +51,9 @@ PROVIDE="virtual/python"
 
 src_prepare() {
 	# Ensure that internal copy of libffi isn't used.
-	rm -fr Modules/_ctypes/libffi*
+	if ! use m68k && ! use sparc-fbsd; then
+		rm -fr Modules/_ctypes/libffi*
+	fi
 
 	if ! tc-is-cross-compiler; then
 		rm "${WORKDIR}/${PV}"/*_all_crosscompile.patch
@@ -155,7 +157,7 @@ src_configure() {
 	fi
 	dbmliborder="${dbmliborder#:}"
 
-	if ! use m68k && ! use mips && ! use sparc-fbsd; then
+	if ! use m68k && ! use sparc-fbsd; then
 		myconf+=" --with-system-ffi"
 	fi
